Extract Building to its own file
